перенес селекторы по папкам
