+++ src/pages/full-offer-page/full-offer-page.tsx
@@ -12,7 +12,8 @@ import UserProfile from '../../components/user-profile/user-profile';
 import ReviewsList from '../../components/reviews-list/reviews-list';
 import { useAppDispatch, useAppSelector } from '../../hooks';
 import { useParams } from 'react-router-dom';
-import { selectAuthorizationStatus, selectFullOffer, selectIsFullOfferLoading, selectIsNearByOffersLoading, selectIsReviewsListLoading, selectNearByOffers, selectReviewList } from '../../store/selectors';
+import { selectFullOffer, selectIsFullOfferLoading, selectIsNearByOffersLoading, selectIsReviewsListLoading, selectNearByOffers, selectReviewList } from '../../store/offers/offers-selectors';
+import { selectAuthorizationStatus} from '../../store/user/user-selector';
 import {fetchOfferInfoByIDAction, fetchOfferReviewListAction, fetchOffesNearAction } from '../../store/api-actions';
 import { useEffect } from 'react';
 import NotFoundPage from '../not-found-page/not-found-page';
